POI card expanded omheen gezet, setheight bij bottomsheet als de route is gestart (gwn beter en mooier gemaakt)

+++ lib/Components/Map/directions_card.dart
@@ -165,7 +165,7 @@ class _DirectionsCardState extends State<DirectionsCard> {
       var totalDistance = distanceToEveryDirection.reduce((a, b) => a + b);
 
       if (disrabceToFinish != totalDistance) {
-        disrabceToFinish = totalDistance;
+        disrabceToFinish = totalDistance.toDouble();
       }
 
       setState(() {
